[#73] fix: 게시물 삭제 전 게시물의 모든 댓글을 삭제하지 않아 발생하는 오류 수정
postId가 유효한지 확인, 유효하다면 모든 댓글 삭제 후 게시물도 삭제하도록 수정함

+++ src/main/java/com/popple/server/domain/board/service/BoardService.java
@@ -91,7 +91,15 @@ public class BoardService {
     }
 
     public void deletePost(Long postId) throws IllegalArgumentException{
+        if(!hasPost(postId)){
+            throw new NoSuchElementException("해당 게시물이 존재하지 않습니다.");
+        }
+        commentRepository.deleteAllByPost_id(postId);
         boardRepository.deleteById(postId);
     }
 
+    public boolean hasPost(Long postId) {
+        Optional<Post> post = boardRepository.findById(postId);
+        return post.isPresent();
+    }
 }
